split NavDBTree
